backend/requirements.txt
fastapi
uvicorn[standard]
cryptography
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-multipart
httpx
pytest-mock
faker
python-jose
pydantic-settings
